+++ ptolemy/math/IntegerArrayMath.java
@@ -687,7 +687,7 @@ public class IntegerArrayMath {
         Complex[] returnValue = new Complex[length];
 
         for (int i = 0; i < length; i++) {
-            returnValue[i] = new Complex((double) array[i], 0.0);
+            returnValue[i] = new Complex(array[i], 0.0);
         }
 
         return returnValue;
@@ -704,7 +704,7 @@ public class IntegerArrayMath {
         double[] returnValue = new double[length];
 
         for (int i = 0; i < length; i++) {
-            returnValue[i] = (double) array[i];
+            returnValue[i] = array[i];
         }
 
         return returnValue;
@@ -722,7 +722,7 @@ public class IntegerArrayMath {
         float[] returnValue = new float[length];
 
         for (int i = 0; i < length; i++) {
-            returnValue[i] = (float) array[i];
+            returnValue[i] = array[i];
         }
 
         return returnValue;
@@ -739,7 +739,7 @@ public class IntegerArrayMath {
         long[] returnValue = new long[length];
 
         for (int i = 0; i < length; i++) {
-            returnValue[i] = (long) array[i];
+            returnValue[i] = array[i];
         }
 
         return returnValue;
