+++ ptolemy/math/IntegerMatrixMath.java
@@ -841,7 +841,7 @@ public class IntegerMatrixMath {
 
         for (int i = 0; i < rows; i++) {
             for (int j = 0; j < columns; j++) {
-                returnValue[i][j] = new Complex((double) matrix[i][j], 0.0);
+                returnValue[i][j] = new Complex(matrix[i][j], 0.0);
             }
         }
 
@@ -861,7 +861,7 @@ public class IntegerMatrixMath {
 
         for (int i = 0; i < rows; i++) {
             for (int j = 0; j < columns; j++) {
-                returnValue[i][j] = (double) matrix[i][j];
+                returnValue[i][j] = matrix[i][j];
             }
         }
 
@@ -881,7 +881,7 @@ public class IntegerMatrixMath {
 
         for (int i = 0; i < rows; i++) {
             for (int j = 0; j < columns; j++) {
-                returnValue[i][j] = (float) matrix[i][j];
+                returnValue[i][j] = matrix[i][j];
             }
         }
 
@@ -901,7 +901,7 @@ public class IntegerMatrixMath {
 
         for (int i = 0; i < rows; i++) {
             for (int j = 0; j < columns; j++) {
-                returnValue[i][j] = (long) matrix[i][j];
+                returnValue[i][j] = matrix[i][j];
             }
         }
