+++ ptolemy/media/javasound/SoundReader.java
@@ -449,7 +449,7 @@ public class SoundReader {
                     result = (result << 8) + (_b[i] & 0xff);
                 }
 
-                _doubleArray[currChannel][currSamp] = ((double) result * maxSampleReciprocal);
+                _doubleArray[currChannel][currSamp] = result * maxSampleReciprocal;
             }
         }
